got rid of ls | xargs cp {} ... ; rm ... for self replication
Files of a transient nature inside of a source repository should not be included in a tar.gz of that repository.  I can't rely on 'git ls' because I support SVN.  I don't want to introduce more dependencies; but I just broke down and did an "ls | grep -v | xargs cp ..." , which is just frowned upon.  I decided that using rsync is OK for this.